add unit test, more renaming, add github actions
